+++ src/OperationalInsights/OperationalInsights/ChangeLog.md
@@ -19,7 +19,8 @@
 -->
 ## Upcoming Release
 * Updated default version for saved searches to be 1. 
-
+* Fixed custom log null regex handling
+  
 ## Version 1.3.1
 * Fixed CustomLog datasource model returned in Get-AzOperationalInsightsDataSource